wip: linux building + compiling
- all projects (except remote console, which is disabled entirely on
  linux) now compile. Almost none of them link though
- Particularly I have no idea what the link issue is with the bc7enc
  library, that might be rough

+++ code/renderer/graphics_api/physical_device.cpp
@@ -268,7 +268,7 @@ PhysicalDevice SelectBestPhysicalDevice( std::vector<PhysicalDevice>& devices, i
     PG_ASSERT( devices.size() > 0 );
     if ( deviceOverride != -1 )
     {
-        PG_ASSERT( deviceOverride < devices.size() );
+        PG_ASSERT( deviceOverride < (int)devices.size() );
         return devices[deviceOverride];
     }
